+++ main-server/src/server.js
@@ -54,7 +54,7 @@ import * as jobWorker from './master-worker-interface/client';
 import * as prometheus from './prometheus';
 import * as telemetryToken from './telemetry/token';
 
-import logger from './logger';
+import logger, { setOptionalErrorLogFn } from './logger';
 
 import TelemetryLogger from './telemetry';
 
@@ -81,6 +81,7 @@ process.on('unhandledRejection', function (reason, p) {
 
 async function initialize() {
   logger.info({ message: 'Initializing server' });
+
   try {
     tendermint.loadSavedData();
 
@@ -94,9 +95,19 @@ async function initialize() {
       prometheus.initialize();
     }
 
-    if (config.telemetryLoggingEnabled) {
-      telemetryDb.initialize();
-      telemetryEventsDb.initialize();
+    if (!config.ndidNode) {
+      if (config.telemetryLoggingEnabled) {
+        telemetryDb.initialize();
+        telemetryEventsDb.initialize();
+
+        setOptionalErrorLogFn((log) => {
+          TelemetryLogger.logProcessLog({
+            nodeId: config.nodeId,
+            process_name: 'main',
+            log,
+          });
+        });
+      }
     }
 
     if (config.ndidNode) {
@@ -108,10 +119,12 @@ async function initialize() {
       tendermint.eventEmitter.once('ready', (status) => resolve(status))
     );
 
-    tendermint.setTelemetryEnabled(
-      (config.mode === MODE.STANDALONE || config.mode === MODE.MASTER) &&
-        config.telemetryLoggingEnabled
-    );
+    if (!config.ndidNode) {
+      tendermint.setTelemetryEnabled(
+        (config.mode === MODE.STANDALONE || config.mode === MODE.MASTER) &&
+          config.telemetryLoggingEnabled
+      );
+    }
 
     await tendermint.connectWS();
     const tendermintStatusOnSync = await tendermintReady;
@@ -204,7 +217,7 @@ async function initialize() {
       await externalCryptoServiceReady;
     }
 
-    if (config.telemetryLoggingEnabled) {
+    if (config.telemetryLoggingEnabled && !config.ndidNode) {
       if (config.mode === MODE.STANDALONE || config.mode === MODE.MASTER) {
         await telemetryToken.initialize({ role });
       }
@@ -287,8 +300,10 @@ async function initialize() {
 
     logger.info({ message: 'Server initialized' });
 
-    if (config.mode === MODE.STANDALONE || config.mode === MODE.MASTER) {
-      TelemetryLogger.logMainVersion({ nodeId: config.nodeId, version });
+    if (!config.ndidNode) {
+      if (config.mode === MODE.STANDALONE || config.mode === MODE.MASTER) {
+        TelemetryLogger.logMainVersion({ nodeId: config.nodeId, version });
+      }
     }
   } catch (error) {
     logger.error({
@@ -355,9 +370,11 @@ async function shutDown() {
 
   await Promise.all([cacheDb.close(), longTermDb.close(), dataDb.close()]);
 
-  if (config.telemetryLoggingEnabled) {
-    await telemetryDb.close();
-    await telemetryEventsDb.close();
+  if (!config.ndidNode) {
+    if (config.telemetryLoggingEnabled) {
+      await telemetryDb.close();
+      await telemetryEventsDb.close();
+    }
   }
 }