+++ main-server/src/telemetry/logger.js
@@ -147,6 +147,17 @@ export default class TelemetryLogger {
     this.abciVersion = abciVersion;
   }
 
+  async logProcessLog({ nodeId, process_name, log }) {
+    if (!this.enable) return;
+
+    await telemetryDb.addProcessLog(nodeId, {
+      node_id: nodeId,
+      process_name,
+      log,
+      // source_timestamp: this.getCurrentTime(),
+    });
+  }
+
   /*
    * logRequestEvent saves the information regarding requestID
    */